+++ packages/thrift-client/package.json
@@ -21,7 +21,7 @@
   "scripts": {
     "clean": "rimraf ./dist ./src/tests/generated",
     "lint": "tslint --fix './src/**/*.ts'",
-    "codegen": "thrift-typescript --target thrift-server --rootDir ./src/tests --sourceDir ./thrift --outDir ./generated",
+    "codegen": "thrift-typescript --strictUnions true --target thrift-server --rootDir ./src/tests --sourceDir ./thrift --outDir ./generated",
     "prebuild": "npm run clean",
     "build": "npm run lint && tsc -p tsconfig.json",
     "prebuild:test": "npm run clean && npm run codegen",
@@ -49,7 +49,7 @@
     "@creditkarma/thrift-server-core": "^0.6.6",
     "@creditkarma/thrift-server-express": "^0.6.6",
     "@creditkarma/thrift-server-hapi": "^0.6.6",
-    "@creditkarma/thrift-typescript": "2.0.0-3",
+    "@creditkarma/thrift-typescript": "2.0.0-5",
     "@types/body-parser": "^1.16.5",
     "@types/code": "^4.0.3",
     "@types/express": "^4.0.37",
