+++ packages/thrift-client/src/tests/client.ts
@@ -13,7 +13,7 @@ import { CoreOptions } from 'request'
 import {
     Calculator,
     Operation,
-    Work,
+    Work_Loose,
 } from './generated/calculator'
 
 import {
@@ -75,7 +75,7 @@ export function createClientServer(sampleRate: number = 0): Promise<net.Server> 
     })
 
     app.get('/calculate', (req: express.Request, res: express.Response): void => {
-        const work: Work = {
+        const work: Work_Loose = {
             num1: req.query.left,
             num2: req.query.right,
             op: symbolToOperation(req.query.op),
@@ -89,7 +89,7 @@ export function createClientServer(sampleRate: number = 0): Promise<net.Server> 
     })
 
     app.get('/calculate-overwrite', (req: express.Request, res: express.Response): void => {
-        const work: Work = {
+        const work: Work_Loose = {
             num1: req.query.left,
             num2: req.query.right,
             op: symbolToOperation(req.query.op),
