+++ packages/thrift-server-hapi/package.json
@@ -44,7 +44,7 @@
   "devDependencies": {
     "@creditkarma/thrift-client": "^0.6.6",
     "@creditkarma/thrift-server-core": "^0.6.6",
-    "@creditkarma/thrift-typescript": "~1.1.4",
+    "@creditkarma/thrift-typescript": "2.0.0-5",
     "@types/code": "^4.0.3",
     "@types/hapi": "^16.1.10",
     "@types/lab": "^11.1.0",
